# repo/api/openapi/v1.yaml
openapi: 3.1.0
info:
  title: KuberCloud ANI Core API
  version: 1.0.0
  description: |
    KuberCloud ANI Core — 基础设施平台 REST API 契约（唯一真实来源）。

    ## 文件范围与边界
    本文件（api/openapi/v1.yaml）仅包含 ANI Core 基础设施资源：
    instances（VM/Container/GPU/Sandbox/BM/K8s集群）、volumes/filesystems/objects/
    vector-stores、networks、auth/secrets/encryption、gpu-inventory/dpu-inventory、
    registry/metering/observability/notifications/audit 等。
    ANI Services 资源（模型仓库/推理/知识库/PaaS）在 api/openapi/services/v1.yaml。

    ## URL 前缀规范
    servers[0].url = "https://{host}/api/v1"（前缀在 spec 中定义，Gateway 不做路径变换）。
    版本升级时新建 api/openapi/v2.yaml，本文件路径保持稳定。

    ## SDK 生成（make gen-core-sdk）
    Go: github.com/kubercloud/ani-go | Python: pip install kubercloud-ani
    TypeScript: @kubercloud/ani-client | Java: com.kubercloud:ani-java

    ## 认证
    Bearer JWT（1小时有效）或 X-API-Key Header。
    tenant_id 从 JWT claims 提取，请求体中的 tenant_id 字段将被忽略。

    ## 错误格式
    `{ code: string, message: string, request_id: string, details?: object }`
    标准错误码：UNAUTHORIZED / FORBIDDEN / NOT_FOUND / CONFLICT / BAD_REQUEST /
    RATE_LIMIT_EXCEEDED / NOT_IMPLEMENTED / INTERNAL_ERROR

    ## 分页
    cursor 分页：`?limit=20&cursor=<token>` → `{ items: [], next_cursor: string | null }`

    ## 异步操作
    202 响应体返回 AsyncTask，Location Header 附任务 URL。
    轮询 GET /api/v1/tasks/{task_id} 或配置 Webhook。
servers:
  - url: https://{host}/api/v1
    variables:
      host:
        default: ani.internal
        description: ANI Gateway 地址（支持IP地址）

...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT AccessToken（有效期1小时，通过 /auth/token 获取）
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: 长期 API Key（在 Console → 设置 → API Key 中创建）

# ── 公共 Schema ───────────────────────────────────────────────────────────────
  schemas:
...
    SecretCreateRequest:
      type: object
      properties:
        idempotency_key: { type: string }
        name:            { type: string }
        type:            { type: string, enum: [opaque, dockerconfigjson, tls] }
        data:
          type: object
          additionalProperties: { type: string }
      required: [idempotency_key, name, data]
    Secret:
      type: object
      properties:
        id:         { type: string }
        tenant_id:  { type: string }
        name:       { type: string }
        type:       { type: string }
        keys:       { type: array, items: { type: string } }
        state:      { type: string, enum: [active, deleted] }
        dev_profile:
          $ref: '#/components/schemas/CoreDevProfileInfo'
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    SecretListResponse:
      type: object
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/Secret" } }
        total: { type: integer }
        next_cursor: { type: string, nullable: true }
    SecretBindingRequest:
      type: object
      properties:
        target_type: { type: string, enum: [instance, k8s-cluster, service] }
        target_id:   { type: string }
        mount_path:  { type: string }
        env_prefix:  { type: string }
      required: [target_type, target_id]
    SecretBinding:
      type: object
      properties:
        id:          { type: string }
        secret_id:   { type: string }
        tenant_id:   { type: string }
        target_type: { type: string }
        target_id:   { type: string }
        mount_path:  { type: string }
        env_prefix:  { type: string }
        state:       { type: string }
        dev_profile:
          $ref: '#/components/schemas/CoreDevProfileInfo'
        created_at:  { type: string, format: date-time }

    # 所有错误响应的统一格式
    ErrorResponse:
      type: object
      required: [code, message, request_id]
      properties:
        code:        { type: string, example: "MODEL_NOT_FOUND" }
        message:     { type: string, example: "模型 qwen2.5-72b 不存在" }
        request_id:  { type: string, example: "req_01j8xabc" }
        details:     { type: object, additionalProperties: true }

    # Cursor 分页（所有列表接口统一使用此格式）
...
    CoreDevProfileInfo:
      type: object
      description: "Core dev/local profile 标记；用于区分本地联调成功与真实 provider 执行成功。"
      required: [mode, provider, real_provider]
      properties:
        mode:          { type: string, enum: [local, real] }
        provider:      { type: string }
        real_provider: { type: boolean }
        reason:        { type: string, nullable: true }
...
  responses:
...
    NotFound:
      description: 资源不存在（code=NOT_FOUND）
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    Conflict:
      description: 资源已存在或冲突（code=CONFLICT）
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []

# ── API 路径定义 ──────────────────────────────────────────────────────────────
# 说明：/demo/* 路径已迁移至 api/openapi/dev.yaml（不参与 SDK 生成）
paths:
...
  /secrets:
    get:
      operationId: listSecrets
      summary: 列出 Secret
      description: 只返回 Secret 元数据和 key 名称，不返回明文值。
      tags: [Secrets]
      x-ani-rbac-scope: "scope:secrets:read"
      parameters:
        - { name: limit,  in: query, schema: { type: integer, minimum: 1, maximum: 100 } }
        - { name: cursor, in: query, schema: { type: string } }
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SecretListResponse"
    post:
      operationId: createSecret
      summary: 创建 Secret
      description: Sprint 5 local dev profile 存储 Secret 明文仅限本地 adapter；API 响应只返回元数据和 key 名称。
      tags: [Secrets]
      x-ani-rbac-scope: "scope:secrets:create"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SecretCreateRequest"
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Secret"
  /secrets/{secret_id}:
    get:
      operationId: getSecret
      summary: 查询 Secret 元数据
      tags: [Secrets]
      x-ani-rbac-scope: "scope:secrets:read"
      parameters:
        - { name: secret_id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Secret"
        "404":
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: deleteSecret
      summary: 删除 Secret
      tags: [Secrets]
      x-ani-rbac-scope: "scope:secrets:delete"
      parameters:
        - { name: secret_id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Secret"
        "404":
          $ref: '#/components/responses/NotFound'
  /secrets/{secret_id}/bindings:
    post:
      operationId: bindSecret
      summary: 绑定 Secret 到工作负载
      description: 当前为 local dev profile 绑定记录；真实 K8s Secret 注入在 provider/controller 边界实现。
      tags: [Secrets]
      x-ani-rbac-scope: "scope:secrets:bind"
      parameters:
        - { name: secret_id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SecretBindingRequest"
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SecretBinding"
        "404":
          $ref: '#/components/responses/NotFound'
        "409":
          $ref: '#/components/responses/Conflict'

# ── Tag 分组（Swagger UI 显示顺序）──────────────────────────────────────────
...